refactor(visio): JitsiMeet sous 300 lignes — éclatement complet (H13)
375 -> 77 l. Logique (script externe, init IFrame API, heartbeat, lifecycle) -> useJitsiMeet.js ; overlays -> JitsiLoadingOverlay/JitsiErrorOverlay. API (props seanceId/jitsiLink/userName/userEmail, emits close/joined/left/error, expose cleanup) inchangee. CSS verbatim. Tests parite + voisins visio verts. Dette : heartbeat conserve inline (parite stricte vs useVisioHeartbeat a Worker) — documentee.
